# NetworkDB performance improvements