@kcrawley 2018-01-11
@fcrisciani will this change potentially solve the limitation on the `/24` cidr block for overlay/vip network? i can link the issue for reference (if needed, but I didn't want to cross-pollinate unless it's related)