@fcrisciani 2018-01-11
@kcrawley actually this is still pretty experimental and I think I will do other changes before having this one ready.
This won't make more scalable ingress because there the bottleneck is the number of IPVS rules that have to be configured into the containers. For that there is other work that is being done that will reduce the complexity from O(n^2) to O(n)